rename `typestats._metadata` -> `typestats.metadata`
