+++ src/typestats/metadata.py
